--- ShapeCrawler/AutoShapes/IField.cs
@@ -1,8 +1,0 @@
-﻿namespace ShapeCrawler;
-
-/// <summary>
-///     Represents a field.
-/// </summary>
-public interface IField
-{
-}
