+++ ShapeCrawler/Texts/IParagraph.cs
