+++ ShapeCrawler/Texts/IPortion.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using DocumentFormat.OpenXml.Packaging;
-using ShapeCrawler.AutoShapes;
 using ShapeCrawler.Exceptions;
 using ShapeCrawler.Extensions;
 using ShapeCrawler.Shared;
