+++ ShapeCrawler/Texts/ITextFrame.cs
