+++ ShapeCrawler/Texts/ITextFrameContainer.cs
